shakesfind-agent/requirements.txt
httpx==0.27.2
requests-cache==1.2.1
extruct==0.16.0
w3lib==2.2.1
selectolax==0.3.24
python-dateutil==2.9.0.post0
dateparser==1.2.0
pytz==2024.1
tldextract==5.1.2
python-dotenv==1.0.1
notion-client==2.2.1
beautifulsoup4==4.12.3
ics==0.7.2
pydantic==2.8.2
PyYAML==6.0.2
fastapi==0.112.2
uvicorn==0.30.5
curl_cffi==0.6.4
cloudscraper==1.2.71
playwright==1.47.0
